requirements/dev.txt
-r base.txt
ipython==8.18.1
django-extensions==3.2.3
django-debug-toolbar~=3.9
pytest-django~=4.5
